fixed warnings and msvc j2k compil

+++ src/scene_manager/encode_isom.c
@@ -274,7 +274,7 @@ static GF_Err gf_sm_import_stream(GF_SceneManager *ctx, GF_ISOFile *mp4, GF_ESD 
 	import.esd = src;
 	import.duration = mux->duration;
 	import.flags = mux->import_flags | GF_IMPORT_FORCE_MPEG4;
-	import.video_fps.num = 1000*mux->frame_rate;
+	import.video_fps.num = (s32) (1000*mux->frame_rate);
 	import.video_fps.den = 1000;
 	import.in_name = szName;
 	import.initial_time_offset = imp_time;
